requirements.txt
wheel
selenium
pytest
openpyxl
matplotlib
pyplot
aspose-cells
chromedriver_autoinstaller
webdriver_manager
PyAutoGUI
numpy
jpype1
pytest-xdist